backend/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pydantic==2.5.3
torch==2.4.1
numpy==1.26.4
psycopg2-binary